Enable half precision

+++ ldm/models/diffusion/ddpm.py
@@ -704,6 +704,7 @@ class LatentDiffusion(DDPM):
 
     @torch.no_grad()
     def decode_first_stage(self, z, predict_cids=False, force_not_quantize=False):
+        z = z.half()
         if predict_cids:
             if z.dim() == 4:
                 z = torch.argmax(z.exp(), dim=1).long()
@@ -889,7 +890,7 @@ class LatentDiffusion(DDPM):
         return [rescale_bbox(b) for b in bboxes]
 
     def apply_model(self, x_noisy, t, cond, return_ids=False):
-
+        x_noisy = x_noisy.half()
         if isinstance(cond, dict):
             # hybrid case, cond is exptected to be a dict
             pass
